iot_device/requirements.txt
paho-mqtt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp